Major rework of the saving system. We now use pickle + some hacks for numpy matrix saving.
This means that all save files are broken, but if the need arise I
will write a conversion script.

On the good side it means that objects no longer need to implement
special methods and that the built state is automatically preserved
across saves.  Also, this is (hopefully) the last rework of this
system.